pyproject.toml
[tool.poetry]
name = "nasake"
version = "0.1.0"
description = ""
authors = ["Devansh3712 <devanshamity@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.104.1"
gunicorn = "^21.2.0"
pydantic-settings = "^2.0.3"
uvicorn = "^0.24.0.post1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
email-validator = "^2.1.0.post1"
bcrypt = "^4.0.1"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
httpx = "^0.25.1"
transformers = "^4.35.2"
tensorflow = "^2.15.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
